use 'figures' instead of buffers, only set uMVMatrix once
demonstrates how I'm starting to understand this stuff better. the MVMatrix is a uniform because it stays constant throughout drawing the figure, because the MVMatrix is only used for the translation (ie, the 'world position') and rotation of the figure as a whole. So, we only need to set it once for each figure. Whereas, the vertex attributes are arrays of vertices which the shaders draw one at a time. Hurray for solidifying some GL knowledge!